feat:공지사항 생성 이후 목록 아닌 상세로 이동하도록
